Handle more change events
Updates display of monitors and current window position
in a range of system events like monitor displays changed,
taskbar auto-hide, primary monitor changed, monitor scaling
displayed, etc.

Made a note in README that DPI change handling is not
covered. It's not clear from MSDN whether that is a manifest
thing or not. Changing scaling resolution in control panel
is detected and handled though, so not sure if DPI is an issue.